+++ packages/expo-ui/src/jetpack-compose/index.ts
@@ -35,6 +35,7 @@ export * from './PullToRefreshBox';
 export * from './RadioButton';
 export * from './Surface';
 export { type TextProps, Text } from './Text';
+export * from './Tooltip';
 
 export * from './AnimatedVisibility';
 export * from './Box';